+++ paintbox/__init__.py
@@ -17,4 +17,3 @@ from .operators import *
 from .sed import *
 from .extlaws import *
 from .likelihoods import *
-    # from .interfaces import *