+++ paintbox/utils/__init__.py
@@ -3,5 +3,5 @@
 # This sub-module is destined for common non-package specific utility
 # functions.
 from .convolve import *
-from .CvD_utils import *
+from .CvD_utils import CvD18
 from .lick import *